requirements.txt
Flask
python-dotenv
requests
emoji
